+++ CHANGELOG.md
@@ -6,6 +6,20 @@ File format follows [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 See cloning and downloading instructions [here](https://www.ivoyager.dev/developers/).
 
+## [v0.0.2] - 2025-03-20
+
+Developed using Godot 4.4.
+
+Breaks older savegame files!
+
+### Changed
+
+* Now handles Godot 4.4 typed dictionaries!
+* Typed all dictionaries in internal code.
+* Changed internal encoding of arrays, dictionaries and objects.
+* Removed unnecessary value indexing.
+* Removed unmaintained debug functions.
+  
 ## v0.0.1 - 2025-03-07
 
 Developed using Godot 4.3. **The next release will have changes for 4.4!**
@@ -14,3 +28,4 @@ Initial alpha release!
 
 This plugin replaces [ivoyager_tree_saver](https://github.com/ivoyager/ivoyager_tree_saver). It has that plugin's functionality plus save/load GUI classes taken out of [ivoyager_core](https://github.com/ivoyager/ivoyager_core).
 
+[v0.0.2]: https://github.com/ivoyager/ivoyager_save/compare/v0.0.1...v0.0.2